requirements.txt
requests~=2.28.1
chardet~=5.1.0
python-dotenv~=0.20.0
selenium~=4.8.2
python-telegram-bot==13.7